+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `on` and `emit` event functions
-- Add `onConnected` and `onDisconnected` lifecycle functions
+- Add `connected` and `disconnected` lifecycle functions
 
 ## [1.0.2] - 2025-06-15